Version 1 Ultimate Complete

+++ Stitch2/Form1.cs
@@ -1,13 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
 using System.Diagnostics;
-using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace Stitch2
@@ -19,31 +15,35 @@ namespace Stitch2
             InitializeComponent();
         }
 
-        private void label1_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void lstDrop_DragEnter(object sender, DragEventArgs e)
+        private void LstDrop_DragEnter(object sender, DragEventArgs e)
         {
             if (e.Data.GetDataPresent(DataFormats.FileDrop)) e.Effect = DragDropEffects.Copy;
         }
 
+
         List<string> rmd_files = new List<string>();
 
-        private void lstDrop_DragDrop(object sender, DragEventArgs e)
+        private void LstDrop_DragDrop(object sender, DragEventArgs e)
         {
-            string[] files = (string[])e.Data.GetData(DataFormats.FileDrop);
+            string[] drag_content = (string[])e.Data.GetData(DataFormats.FileDrop);
 
-            foreach (string file_str in files)
+            //For each of the dropped content...Get the rmd files
+            foreach (string file_str in drag_content)
             {
                 foreach(string file in GetFilesList(file_str))
                 {
-                    MessageBox.Show(file);
+                    String file_name = Path.GetFileNameWithoutExtension(file);
+        
+                    if (!rmd_files.Contains(file))
+                    {
+                        rmd_files.Add(file);
+                        lstDrop.Items.Add(file_name);
+                    }
                 }
             }
         }
-
+    
+        //Returns list of all rmd files in a dir
         private List<String> GetFilesList(String dir)
         {
             List<String> results = new List<string>();
@@ -65,7 +65,7 @@ namespace Stitch2
         }
                 
 
-        private void btnStitch_Click(object sender, EventArgs e)
+        private void BtnStitch_Click(object sender, EventArgs e)
         {
             //Save the stuff to text file 
             String desktop_path = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
@@ -74,16 +74,14 @@ namespace Stitch2
             TextWriter tw = new StreamWriter(fileName);
             foreach(String s in rmd_files)
             {
-                tw.WriteLine(s);
+                tw.WriteLine(s.Replace("\\", @"/"));
             }
             tw.Close();
 
-            //Run the program on the text file 
+            //Run the program on the text file
             Process p = new Process();
-            //p.StartInfo.UseShellExecute = false;
-            //p.StartInfo.RedirectStandardOutput = true;
-            p.StartInfo.FileName = "cmd.exe";
-            p.StartInfo.Arguments = "powershell -windowstyle hidden -executionpolicy remotesigned -File  Stitch.ps1 -f " + fileName;
+            p.StartInfo.FileName = "powershell";
+            p.StartInfo.Arguments = " -executionpolicy remotesigned -File  Stitch.ps1 -f " + fileName;
             p.Start();
 
             p.WaitForExit();
@@ -93,30 +91,15 @@ namespace Stitch2
             List<String> failed = new List<string>();
 
             IEnumerable<string> lines = File.ReadLines(fileName + "-fail.txt");
-            foreach(string line in lines)
-            {
-                failed.Add(line);
-            }
-
-            if (File.Exists(fileName))
-            {
-                File.Delete(fileName);
-            }
-
-            if (File.Exists(fileName + "-fail.txt")){
-                File.Delete(fileName + "-fail.txt");
-            }
+            foreach (string line in lines) failed.Add(line);
 
             IEnumerable<string> lines2 = File.ReadLines(fileName + "-succeed.txt");
-            foreach (string line in lines2)
-            {
-                passed.Add(line);
-            }
+            foreach (string line in lines2) passed.Add(line);
 
-            if (File.Exists(fileName + "-succeed.txt"))
-            {
-                File.Delete(fileName + "-succeed.txt");
-            }
+
+            if (File.Exists(fileName)) File.Delete(fileName);
+            if (File.Exists(fileName + "-fail.txt")) File.Delete(fileName + "-fail.txt");
+            if (File.Exists(fileName + "-succeed.txt")) File.Delete(fileName + "-succeed.txt");
             //Send Data to the other form
             Report b = new Report();
             b.LoadDetails(passed, failed);
@@ -126,6 +109,12 @@ namespace Stitch2
         private void Form1_Load(object sender, EventArgs e)
         {
             this.CenterToScreen();
+        }
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            rmd_files.Clear();
+            lstDrop.Items.Clear();
         }
     }
 }
